Modify breeding to account for fitness of NaN
After mutation in the breeding function, the max_fitness is initialised
to the first individual in the new_generation. If the fitness (returned
from calcFitness) evaluates to -nan then the max_fitness is never
updated. This is because (x > -nan) evaluates to false. Note that
(x < -nan) also evaluates to false.